requirements.txt
aiofiles==25.1.0
anyio==4.13.0
boto3==1.43.0
botocore==1.43.0
certifi==2026.4.22
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jmespath==1.1.0
numpy==2.4.4
pillow==12.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
s3transfer==0.17.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
